+++ src/routes/audits.tsx
@@ -1,0 +1,3 @@
+import { createFileRoute } from '@tanstack/react-router'
+import { AuditIndex } from '../components/AuditIndex'
+export const Route = createFileRoute('/audits')({ component: AuditIndex })